Linear program

+++ Headers/helpers/string.h
@@ -3,6 +3,6 @@
 
 /* Usage functions */
 
-int str_to_int(const char * s);
+unsigned short str_to_int(const char * s);
 
 #endif //SNAKE_STRING_H
